+++ lib/ig_markets/cli/commands/positions_command.rb
@@ -130,7 +130,7 @@ module IGMarkets
         currency_totals = positions.group_by(&:currency).map do |currency, subset|
           total = subset.map(&:profit_loss).reduce :+
 
-          Format.currency(total, currency).colorize(total < 0 ? :red : :green)
+          Format.colored_currency total, currency
         end
 
         puts "\nTotal profit/loss: #{currency_totals.join ', '}"